[TSK] Contects Lists [NEW] Crindo layout para os contatos [NEW] Ceparando os widget em diretorios [REF] Tirando o dashboard do main e criando um arquivo para ele
