+++ lib/screens/contacts_list.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+class ContectsList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Contects'),
+      ),
+      body: ListView(
+        children: <Widget>[
+          Card(
+            child: ListTile(
+              title: Text(
+                'teste',
+                style: TextStyle(fontSize: 24.0),
+              ),
+              subtitle: Text(
+                '1000',
+                style: TextStyle(fontSize: 16.0),
+              ),
+            ),
+          )
+        ],
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {},
+        child: Icon(Icons.add),
+      ),
+    );
+  }
+}
